LPS-86455 Move event listener interface to repository package
